[TERK-150/Feat] 상품 재고 validation api 추가

+++ src/main/java/com/ssg/sausageitemapi/item/dto/response/ItemFindResponse.java
@@ -29,7 +29,7 @@ public class ItemFindResponse {
     private Integer itemAmt;
 
     @Schema(description = "상품 수량")
-    private Long itemInvQty;
+    private Integer itemInvQty;
 
     @Schema(description = "멤버 ID")
     private String itemImgUrl;
